Large update that corrects a lot of small errors and tries to simplify the way we extract xml data. Update the docs to reflect current state.

+++ pywind/ofgem/search.py
@@ -100,7 +100,7 @@ class CertificateSearch(object):
         """
         if not isinstance(yearmonth, int):
             yearmonth = int(yearmonth)
-        year = yearmonth / 100
+        year = int(yearmonth / 100)
         if self._set_year(year) is False:
             return False
         return self._set_month(yearmonth % year)
